backend/pyproject.toml
[tool.poetry]
name = "contentgen-backend"
version = "0.1.0"
description = "FastAPI backend for ContentGen"
authors = ["you <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
httpx = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
psycopg = {extras = ["binary"], version = "^3.2"}
databases = {extras = ["postgresql"], version = "^0.9.0"}
pydantic = "^2.9"
python-dotenv = "^1.0.0"
requests = "^2.32.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-asyncio = "^0.24.0"
ruff = "^0.8.0"
mypy = "^1.13"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
